+++ src/InternalVarDumper.php
@@ -53,15 +53,6 @@ class InternalVarDumper implements VarDumperInterface
         return $result;
     }
 
-    public function getDump($var)
-    {
-        ++$this->shift;
-        $result = $this->dumpAsString($var);
-        --$this->shift;
-
-        return $result;
-    }
-
     protected function dumpPlaceInCode($number)
     {
         $options = \version_compare(PHP_VERSION, '5.3.6') >= 0 ? DEBUG_BACKTRACE_IGNORE_ARGS : false;